refactor: extract layer actions mainline
